wt88047: camera: Partial update HAL post processing
Use rendang camera source as reference

+++ camera/QCamera2/HAL/QCameraMem.cpp
@@ -178,7 +178,7 @@ int QCameraMemory::getSize(int index) const
  *
  * RETURN     : number of buffers allocated
  *==========================================================================*/
-int QCameraMemory::getCnt() const
+uint8_t QCameraMemory::getCnt() const
 {
     return mBufferCount;
 }
